cost-efficient-llm/requirements.txt
openai>=1.0.0
anthropic>=0.5.0
transformers>=4.30.0
torch>=2.0.0
python-dotenv>=0.19.0
groq>=0.4.0
aiohttp>=3.8.0